Add directories src/utils/collector, src/utils/loader
