+++ src/components/block/BlockLoader.ts
@@ -18,7 +18,7 @@
  *
  */
 
-import {EntityLoader} from "@/utils/EntityLoader";
+import {EntityLoader} from "@/utils/loader/EntityLoader";
 import {Block} from "@/schemas/HederaSchemas";
 import {computed, ComputedRef, Ref} from "vue";
 import axios, {AxiosResponse} from "axios";
