+++ src/components/values/TokenAmount.vue
@@ -40,7 +40,7 @@
 import {computed, defineComponent, inject, onMounted, ref, watch} from "vue";
 import {AxiosResponse} from "axios";
 import {TokenInfo} from "@/schemas/HederaSchemas";
-import {TokenInfoCollector} from "@/utils/TokenInfoCollector";
+import {TokenInfoCollector} from "@/utils/collector/TokenInfoCollector";
 import TokenExtra from "@/components/values/TokenExtra.vue";
 import {initialLoadingKey} from "@/AppKeys";
